+++ examples/next-openai/app/use-chat-persistence/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
-import { createChat } from '../api/use-chat-persistence/chat-store';
+import { createChat } from '@util/chat-store';
 
 export default async function ChatPage() {
   const id = await createChat();