+++ packages/ai/core/generate-text/stream-text.ts
@@ -145,7 +145,8 @@ export function streamText<
   experimental_continueSteps: continueSteps = false,
   experimental_telemetry: telemetry,
   experimental_providerMetadata: providerMetadata,
-  experimental_toolCallStreaming: toolCallStreaming = false,
+  experimental_toolCallStreaming = false,
+  toolCallStreaming = experimental_toolCallStreaming,
   experimental_activeTools: activeTools,
   experimental_repairToolCall: repairToolCall,
   experimental_transform: transform,
@@ -227,6 +228,11 @@ A function that attempts to repair a tool call that failed to parse.
     /**
 Enable streaming of tool call deltas as they are generated. Disabled by default.
      */
+    toolCallStreaming?: boolean;
+
+    /**
+@deprecated Use `toolCallStreaming` instead.
+     */
     experimental_toolCallStreaming?: boolean;
 
     /**
@@ -247,6 +253,7 @@ Callback that is called for each chunk of the stream. The stream processing will
         {
           type:
             | 'text-delta'
+            | 'reasoning'
             | 'tool-call'
             | 'tool-call-streaming-start'
             | 'tool-call-delta'
@@ -422,6 +429,9 @@ class DefaultStreamTextResult<
   private readonly textPromise = new DelayedPromise<
     Awaited<StreamTextResult<TOOLS, PARTIAL_OUTPUT>['text']>
   >();
+  private readonly reasoningPromise = new DelayedPromise<
+    Awaited<StreamTextResult<TOOLS, PARTIAL_OUTPUT>['reasoning']>
+  >();
   private readonly toolCallsPromise = new DelayedPromise<
     Awaited<StreamTextResult<TOOLS, PARTIAL_OUTPUT>['toolCalls']>
   >();
@@ -503,6 +513,7 @@ class DefaultStreamTextResult<
             {
               type:
                 | 'text-delta'
+                | 'reasoning'
                 | 'tool-call'
                 | 'tool-call-streaming-start'
                 | 'tool-call-delta'
@@ -541,6 +552,7 @@ class DefaultStreamTextResult<
     let recordedStepText = '';
     let recordedContinuationText = '';
     let recordedFullText = '';
+    let recordedReasoningText: string | undefined = undefined;
     const recordedResponse: LanguageModelResponseMetadata & {
       messages: Array<ResponseMessage>;
     } = {
@@ -568,6 +580,7 @@ class DefaultStreamTextResult<
 
         if (
           part.type === 'text-delta' ||
+          part.type === 'reasoning' ||
           part.type === 'tool-call' ||
           part.type === 'tool-result' ||
           part.type === 'tool-call-streaming-start' ||
@@ -580,6 +593,11 @@ class DefaultStreamTextResult<
           recordedStepText += part.textDelta;
           recordedContinuationText += part.textDelta;
           recordedFullText += part.textDelta;
+        }
+
+        if (part.type === 'reasoning') {
+          recordedReasoningText =
+            (recordedReasoningText ?? '') + part.textDelta;
         }
 
         if (part.type === 'tool-call') {
@@ -625,6 +643,7 @@ class DefaultStreamTextResult<
           const currentStepResult: StepResult<TOOLS> = {
             stepType,
             text: recordedStepText,
+            reasoning: recordedReasoningText,
             toolCalls: recordedToolCalls,
             toolResults: recordedToolResults,
             finishReason: part.finishReason,
@@ -700,6 +719,7 @@ class DefaultStreamTextResult<
 
           // aggregate results:
           self.textPromise.resolve(recordedFullText);
+          self.reasoningPromise.resolve(recordedReasoningText);
           self.stepsPromise.resolve(recordedSteps);
 
           // call onFinish callback:
@@ -708,6 +728,7 @@ class DefaultStreamTextResult<
             logprobs: undefined,
             usage,
             text: recordedFullText,
+            reasoning: recordedReasoningText,
             toolCalls: lastStep.toolCalls,
             toolResults: lastStep.toolResults,
             request: lastStep.request ?? {},
@@ -940,6 +961,7 @@ class DefaultStreamTextResult<
           let stepProviderMetadata: ProviderMetadata | undefined;
           let stepFirstChunk = true;
           let stepText = '';
+          let stepReasoning = '';
           let fullStepText = stepType === 'continue' ? previousStepText : '';
           let stepLogProbs: LogProbs | undefined;
           let stepResponse: { id: string; timestamp: Date; modelId: string } = {
@@ -1042,7 +1064,12 @@ class DefaultStreamTextResult<
                       } else {
                         await publishTextChunk({ controller, chunk });
                       }
+                      break;
+                    }
 
+                    case 'reasoning': {
+                      controller.enqueue(chunk);
+                      stepReasoning += chunk.textDelta;
                       break;
                     }
 
@@ -1325,6 +1352,10 @@ class DefaultStreamTextResult<
     return this.textPromise.value;
   }
 
+  get reasoning() {
+    return this.reasoningPromise.value;
+  }
+
   get toolCalls() {
     return this.toolCallsPromise.value;
   }
@@ -1442,6 +1473,13 @@ However, the LLM results are expected to be small enough to not cause issues.
         switch (chunkType) {
           case 'text-delta': {
             controller.enqueue(formatDataStreamPart('text', chunk.textDelta));
+            break;
+          }
+
+          case 'reasoning': {
+            controller.enqueue(
+              formatDataStreamPart('reasoning', chunk.textDelta),
+            );
             break;
           }