+++ packages/anthropic/src/anthropic-messages-language-model.ts
@@ -109,7 +109,6 @@ export class AnthropicMessagesLanguageModel implements LanguageModelV1 {
     const { prompt: messagesPrompt, betas: messagesBetas } =
       convertToAnthropicMessagesPrompt({
         prompt,
-        cacheControl: this.settings.cacheControl ?? false,
       });
 
     const baseArgs = {
@@ -178,10 +177,6 @@ export class AnthropicMessagesLanguageModel implements LanguageModelV1 {
     betas: Set<string>;
     headers: Record<string, string | undefined> | undefined;
   }) {
-    if (this.settings.cacheControl) {
-      betas.add('prompt-caching-2024-07-31');
-    }
-
     return combineHeaders(
       await resolve(this.config.headers),
       betas.size > 0 ? { 'anthropic-beta': Array.from(betas).join(',') } : {},
@@ -258,17 +253,13 @@ export class AnthropicMessagesLanguageModel implements LanguageModelV1 {
         modelId: response.model ?? undefined,
       },
       warnings,
-      providerMetadata:
-        this.settings.cacheControl === true
-          ? {
-              anthropic: {
-                cacheCreationInputTokens:
-                  response.usage.cache_creation_input_tokens ?? null,
-                cacheReadInputTokens:
-                  response.usage.cache_read_input_tokens ?? null,
-              },
-            }
-          : undefined,
+      providerMetadata: {
+        anthropic: {
+          cacheCreationInputTokens:
+            response.usage.cache_creation_input_tokens ?? null,
+          cacheReadInputTokens: response.usage.cache_read_input_tokens ?? null,
+        },
+      },
       request: { body: JSON.stringify(args) },
     };
   }
@@ -417,16 +408,14 @@ export class AnthropicMessagesLanguageModel implements LanguageModelV1 {
                 usage.promptTokens = value.message.usage.input_tokens;
                 usage.completionTokens = value.message.usage.output_tokens;
 
-                if (self.settings.cacheControl === true) {
-                  providerMetadata = {
-                    anthropic: {
-                      cacheCreationInputTokens:
-                        value.message.usage.cache_creation_input_tokens ?? null,
-                      cacheReadInputTokens:
-                        value.message.usage.cache_read_input_tokens ?? null,
-                    },
-                  };
-                }
+                providerMetadata = {
+                  anthropic: {
+                    cacheCreationInputTokens:
+                      value.message.usage.cache_creation_input_tokens ?? null,
+                    cacheReadInputTokens:
+                      value.message.usage.cache_read_input_tokens ?? null,
+                  },
+                };
 
                 controller.enqueue({
                   type: 'response-metadata',