+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,34 @@
 
 ### Patch Changes
 
+- Updated dependencies [3c5fafa]
+- Updated dependencies [3a58a2e]
+  - ai@4.1.2
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0a699f1]
+  - ai@4.1.1
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [62ba5ad]
+  - ai@4.1.0
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.0.41
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [33592d2]
   - ai@4.0.40