+++ packages/react/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @ai-sdk/react
 
+## 1.1.2
+
+### Patch Changes
+
+- 6f4d063: fix (ai/react): cache addToolResult in useChat
+- Updated dependencies [ed012d2]
+  - @ai-sdk/provider-utils@2.1.2
+  - @ai-sdk/ui-utils@1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [e7a9ec9]
+- Updated dependencies [0a699f1]
+  - @ai-sdk/ui-utils@1.1.1
+  - @ai-sdk/provider-utils@2.1.1
+
+## 1.1.0
+
+### Minor Changes
+
+- 62ba5ad: release: AI SDK 4.1
+
+### Patch Changes
+
+- Updated dependencies [62ba5ad]
+  - @ai-sdk/provider-utils@2.1.0
+  - @ai-sdk/ui-utils@1.1.0
+
+## 1.0.14
+
+### Patch Changes
+
+- 44f04d5: feat (ai/react): expose chat id in experimental_prepareRequestBody
+
 ## 1.0.13
 
 ### Patch Changes